Move JSONNode into UDFNode as only it uses it
