+++ pipeline/json.go
@@ -5,10 +5,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"sort"
-	"strconv"
 	"time"
 
-	"github.com/influxdata/influxdb/influxql"
 	"github.com/influxdata/kapacitor/tick/ast"
 )
 
@@ -116,195 +114,6 @@ func (p *PipelineSorter) visit(node ID, childrenOf Graph) error {
 	p.permanent[node] = true
 	p.sorted = append([]ID{node}, p.sorted...)
 	return nil
-}
-
-// JSONNode contains all fields associated with a node.  `typeOf`
-// is used to determine which type of node this is.
-type JSONNode map[string]interface{}
-
-// NewJSONNode decodes JSON bytes into a JSONNode
-func NewJSONNode(data []byte) (JSONNode, error) {
-	dec := json.NewDecoder(bytes.NewReader(data))
-	dec.UseNumber()
-	var input JSONNode
-	err := dec.Decode(&input)
-	return input, err
-}
-
-// CheckTypeOf tests that the typeOf field is correctly set to typ.
-func (j JSONNode) CheckTypeOf(typ string) error {
-	t, ok := j[NodeTypeOf]
-	if !ok {
-		return fmt.Errorf("missing typeOf field")
-	}
-
-	if t != typ {
-		return fmt.Errorf("error unmarshaling node type %s; received %s", typ, t)
-	}
-	return nil
-}
-
-// SetType adds the Node type information
-func (j JSONNode) SetType(typ string) JSONNode {
-	j[NodeTypeOf] = typ
-	return j
-}
-
-// SetID adds the Node ID information
-func (j JSONNode) SetID(id ID) JSONNode {
-	j[NodeID] = fmt.Sprintf("%d", id)
-	return j
-}
-
-// Set adds the key/value to the JSONNode
-func (j JSONNode) Set(key string, value interface{}) JSONNode {
-	j[key] = value
-	return j
-}
-
-// SetDuration adds key to the JSONNode but formats the duration in InfluxQL style
-func (j JSONNode) SetDuration(key string, value time.Duration) JSONNode {
-	return j.Set(key, influxql.FormatDuration(value))
-}
-
-// Field returns expected field or error if field doesn't exist
-func (j JSONNode) Field(field string) (interface{}, error) {
-	fld, ok := j[field]
-	if !ok {
-		return nil, fmt.Errorf("missing expected field %s", field)
-	}
-	return fld, nil
-}
-
-// String reads the field for a string value
-func (j JSONNode) String(field string) (string, error) {
-	s, err := j.Field(field)
-	if err != nil {
-		return "", err
-	}
-
-	str, ok := s.(string)
-	if !ok {
-		return "", fmt.Errorf("field %s is not a string value but is %T", field, s)
-	}
-	return str, nil
-}
-
-// Int64 reads the field for a int64 value
-func (j JSONNode) Int64(field string) (int64, error) {
-	n, err := j.Field(field)
-	if err != nil {
-		return 0, err
-	}
-
-	jnum, ok := n.(json.Number)
-	if ok {
-		return jnum.Int64()
-	}
-	num, ok := n.(int64)
-	if ok {
-		return num, nil
-	}
-	return 0, fmt.Errorf("field %s is not an integer value but is %T", field, n)
-}
-
-// Float64 reads the field for a float64 value
-func (j JSONNode) Float64(field string) (float64, error) {
-	n, err := j.Field(field)
-	if err != nil {
-		return 0, err
-	}
-
-	jnum, ok := n.(json.Number)
-	if ok {
-		return jnum.Float64()
-	}
-	num, ok := n.(float64)
-	if ok {
-		return num, nil
-	}
-	return 0, fmt.Errorf("field %s is not a floating point value but is %T", field, n)
-}
-
-// Strings reads the field an array of strings
-func (j JSONNode) Strings(field string) ([]string, error) {
-	s, err := j.Field(field)
-	if err != nil {
-		return nil, err
-	}
-
-	strs, ok := s.([]string)
-	if !ok {
-		return nil, fmt.Errorf("field %s is not an array of strings but is %T", field, s)
-	}
-	return strs, nil
-}
-
-// Duration reads the field and assumes the string is in InfluxQL Duration format.
-func (j JSONNode) Duration(field string) (time.Duration, error) {
-	d, err := j.Field(field)
-	if err != nil {
-		return 0, err
-	}
-
-	dur, ok := d.(string)
-	if !ok {
-		return 0, fmt.Errorf("field %s is not a string duration value but is %T", field, d)
-	}
-
-	return influxql.ParseDuration(dur)
-}
-
-// Bool reads the field for a boolean value
-func (j JSONNode) Bool(field string) (bool, error) {
-	b, err := j.Field(field)
-	if err != nil {
-		return false, err
-	}
-
-	boolean, ok := b.(bool)
-	if !ok {
-		return false, fmt.Errorf("field %s is not a bool value but is %T", field, b)
-	}
-	return boolean, nil
-}
-
-// Lambda reads the field as an ast.LambdaNode
-func (j JSONNode) Lambda(field string) (*ast.LambdaNode, error) {
-	n, err := j.Field(field)
-	if err != nil {
-		return nil, err
-	}
-
-	if n == nil {
-		return nil, nil
-	}
-
-	lamb, ok := n.(map[string]interface{})
-	if !ok {
-		return nil, fmt.Errorf("field %s is not a lambda expression but is %T", field, n)
-	}
-
-	lambda := &ast.LambdaNode{}
-	err = lambda.Unmarshal(lamb)
-	return lambda, err
-}
-
-// Type returns the Node type.  This name can be used
-// as the chain function name for the parent node.
-func (j JSONNode) Type() (string, error) {
-	return j.String(NodeTypeOf)
-}
-
-// ID returns the unique ID for this node.  This ID is used
-// as the id of the parent and children in the Edges structure.
-func (j JSONNode) ID() (ID, error) {
-	i, err := j.String(NodeID)
-	if err != nil {
-		return 0, err
-	}
-	id, err := strconv.Atoi(i)
-	return ID(id), err
 }
 
 func (p *Pipeline) MarshalJSON() ([]byte, error) {
